chore(test): move QueueMetricsPublisherTest from orca
This used to live in AtlasQueueMonitor in orca, but this function is now in QueueMetricsPublisher.

As part of the move, I changed the tests from using mock counters and a mock registry to the recommended Spectator pattern for testing using a DefaultRegistry. This is more compact we no longer need to reset the mocks after each test.